headerコンポーネントを一つにまとめた
haederコンポーネントの中にNavコンポーネントを作って分けてた
hooksにロジックを書いてまとめた
