--- src/components/Header/Nav/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Nav";
